Move Percentage from length.rs to mod.rs
