+++ tests/unit/style/attr.rs
@@ -4,8 +4,7 @@
 
 use app_units::Au;
 use style::attr::{AttrValue, LengthOrPercentageOrAuto, parse_length};
-use style::values::computed::CalcLengthOrPercentage;
-use style::values::computed::length::Percentage;
+use style::values::computed::{CalcLengthOrPercentage, Percentage};
 
 #[test]
 fn test_length_calc() {
